+++ src/models.rs
@@ -17,9 +17,6 @@ pub use st7789vw::*;
 pub trait Model {
     type ColorFormat: RgbColor;
 
-    /// Common model constructor
-    fn new() -> Self;
-
     /// Initializes the display for this model
     /// and returns the value of MADCTL set by init
     fn init<RST, DELAY, DI>(